+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.3] - 
 ### Added
+- Local Storage Added (Cookie and LocalStorage can be chosen by AuthProvider Prop `authStorageType`)
 - Documentation [https://authkit.arkadip.co](https://authkit.arkadip.co)
 - PrivateRoute now supports render and component props
 - Auth Token Type added on Sign In. Default `Bearer`